+++ lib/controller.js
@@ -1,117 +1,124 @@
-const http = require('http');
-const https = require('https');
+const _ = require('lodash');
+const Boom = require('boom');
+const assert = require('assert');
+const JWT = require('jsonwebtoken');
+const superagent = require('superagent');
 
-const BEGIN_KEY = '-----BEGIN RSA PUBLIC KEY-----\n';
-const END_KEY = '\n-----END RSA PUBLIC KEY-----\n';
+const KeycloakCerts = require('./getPublicKey');
 
-module.exports = function KeycloakPublicKeyFetcher(url, realm) {
-  const certsUrl = realm ? `${url}/auth/realms/${realm}/protocol/openid-connect/certs` : url;
+/**
+ * 
+ * @param {Object}  server                  Hapi server we are attachhing plugin to
+ * @param {Object}  options                 Configuration options passed
+ * @param {String}  options.issuer          Keycloak's issuer
+ * @param {String}  options.baseUrl         Keycloak base URL
+ * @param {String}  options.realm           Keycloak realm to validate against
+ * @param {Boolean} [options.introspect]    If false, use keycloak's token instrospect
+ * @param {String}  [options.clientId]      Required if using token introspection
+ * @param {String}  [options.clientSecret]  Required if using token introspection
+ */
+function controller(server, options) {
+  if (_.isNil(options.introspect)) {
+    options.usePublicKey = false;
+  }
+  assert(options, 'options are required for keycloak-jwt scheme');
+  assert(options.issuer, 'keycloak `issuer` is required');
+  assert(options.baseUrl, 'keycloak `baseUrl` is required');
+  assert(options.realm, 'keycloak `realm` is required');
+  assert(
+    options.introspect ? options.clientId && options.clientSecret : true,
+    'keycloak `clientId` and `clientSecret` are required to perform token introspect'
+  );
+
+  /**
+   * Verify the keycloak token
+   * @param {String}  token Keycloak token from request
+   * @param {String}  key   Keycloak public key
+   */
+  const verify = function verify(token, key) {
+    let data;
+    try {
+      data = JWT.verify(token, key);
+      return Promise.resolve(true);
+    } catch (e) {
+      return Promise.resolve(false);
+    }
+  }
+
+  options.certFunc = new KeycloakCerts(options.baseUrl, options.realm);
+
   return {
-    fetch: kid => fetch(certsUrl, kid)
-  };
-};
-
-function fetch(url, kid) {
-  return getJson(url)
-    .then((response) => {
-      const key = getKey(response, kid);
-      if (!key) {
-        throw new Error(`Can't find key for kid "${kid}" in response.`);
+    /**
+    * Authenticator for the scheme.
+    * If unauthorize, reply with Boom null message and auth scheme to try other strategies
+    * @param {Object} request  Hapi route handler
+    * @param {Object} reply    Hapi reply interface
+    */
+    authenticate: function(request, reply) {
+      let token = request.headers.authorization;
+      token = token ? token.replace(/Bearer/gi, '').replace(/ /g, '') : null;
+      if (_.isNil(token)) {
+        token = token ? token.replace(/bearer/gi, '').replace(/ /g, '') : null;
       }
-      verify(key);
-      return Promise.resolve(getPublicKey(key.n, key.e));
-    })
-    .catch(error => Promise.reject(err));
-}
-
-function getJson(url) {
-  return new Promise((resolve, reject) => {
-    const agent = url.startsWith('https') ? https : http;
-    agent.get(url, (res) => {
-      if (!valid(res)) {
-        res.resume();
-        reject(new Error(`Status: ${res.statusCode}, Content-type: ${res.headers['content-type']}`));
-      }
-      parse(res)
-        .then(result => resolve(result))
-        .catch(error => reject(error));
-    }).on('error', (e) => { reject(e); });
-  });
-}
-
-function valid(response) {
-  return response.statusCode === 200 && response.headers['content-type'] === 'application/json';
-}
-
-function parse(response) {
-  return new Promise((resolve, reject) => {
-    let rawData = '';
-    response.setEncoding('utf8');
-    response.on('data', (chunk) => { rawData += chunk; });
-    response.on('end', () => {
+      let decoded;
       try {
-        const parsedData = JSON.parse(rawData);
-        resolve(parsedData);
+        decoded = JWT.decode(token, { complete: true });
       } catch (e) {
-        reject(e);
+        return reply(Boom.unauthorized(null, 'keycloak-jwt'));
       }
-    });
-  });
+
+      if (options.introspect) {
+        const basicToken = Buffer.from(`${options.clientId}:${options.clientSecret}`).toString('base64');
+        const url = `${options.issuer}/protocol/openid-connect/token/introspect`;
+        const headers = {
+          Authorization: `Basic ${basicToken}`,
+          connection: 'keep-alive',
+          'content-type': 'application/x-www-form-urlencoded'
+        };
+        const body = { token: token }
+        return superagent('POST', url)
+          .set(headers)
+          .send(body)
+          .then((response) => {
+            const { body } = response;
+            if (!body.active) {
+              return reply(Boom.unauthorized(null, 'keycloak-jwt'));
+            }
+            return reply.continue({
+              credentials: decoded.payload,
+              artifacts: token
+            });
+          })
+          .catch(err => reply(Boom.unauthorized(null, 'keycloak-jwt')));
+      }
+    
+      let promise = Promise.resolve([]);
+      if (_.isNil(options.publicKey)) {
+        const kid = decoded.header.kid;
+    
+        promise = options.certFunc.fetch(kid)
+          .then((key) => {
+            options.publicKey = key;
+            return verify(token, key);
+          })
+          .catch(err => reply(Boom.unauthorized(null, 'keycloak-jwt')));
+      } else {
+        promise = verify(token, options.publicKey);
+      }
+    
+      return promise
+        .then((response) => {
+          const unauthorized = !(typeof response === 'boolean' && response === true);
+          if (!unauthorized) {
+            return reply.continue({
+              credentials: decoded.payload,
+              artifacts: token
+            });
+          }
+          return reply(Boom.unauthorized(null, 'keycloak-jwt'));
+        });
+    }
+  };
 }
 
-function getKey(response, kid) {
-  return Object.hasOwnProperty.call(response, 'keys')
-    ? response.keys.find(k => k.kid === kid)
-    : undefined;
-}
-
-function verify(key) {
-  if (!(key.n && key.e)) {
-    throw new Error('Can\'t find modulus or exponent in key.');
-  }
-  if (key.kty !== 'RSA') {
-    throw new Error('Key type (kty) must be RSA.');
-  }
-  if (key.alg !== 'RS256') {
-    throw new Error('Algorithm (alg) must be RS256.');
-  }
-}
-
-// Based on tracker1's node-rsa-pem-from-mod-exp module.
-// See https://github.com/tracker1/node-rsa-pem-from-mod-exp
-function getPublicKey(modulus, exponent) {
-  const mod = convertToHex(modulus);
-  const exp = convertToHex(exponent);
-  const encModLen = encodeLenght(mod.length / 2);
-  const encExpLen = encodeLenght(exp.length / 2);
-  const part = [mod, exp, encModLen, encExpLen].map(n => n.length / 2).reduce((a, b) => a + b);
-  const bufferSource = `30${encodeLenght(part + 2)}02${encModLen}${mod}02${encExpLen}${exp}`;
-  const pubkey = Buffer.from(bufferSource, 'hex').toString('base64');
-  return BEGIN_KEY + pubkey.match(/.{1,64}/g).join('\n') + END_KEY;
-}
-
-function convertToHex(str) {
-  const hex = Buffer.from(str, 'base64').toString('hex');
-  return hex[0] < '0' || hex[0] > '7'
-    ? `00${hex}`
-    : hex;
-}
-
-function encodeLenght(n) {
-  return n <= 127
-    ? toHex(n)
-    : toLongHex(n);
-}
-
-function toLongHex(number) {
-  const str = toHex(number);
-  const lengthByteLength = 128 + (str.length / 2);
-  return toHex(lengthByteLength) + str;
-}
-
-function toHex(number) {
-  const str = number.toString(16);
-  return (str.length % 2)
-    ? `0${str}`
-    : str;
-}
+module.exports = controller;
